feat: configurable Layout in demo
