+++ apps/demo/components/Layout.tsx
@@ -1,13 +1,12 @@
 // External libraries
 import Image from "next/image";
 import Link from "next/link";
+import { useRouter } from "next/router";
 
 import { FC, ReactNode, useState } from "react";
 
 // SK Components
 import {
-  Button,
-  FAB,
   MaterialIcon,
   NavBar,
   NavBarItem,
@@ -15,17 +14,28 @@ import {
   NavDrawerItem,
   NavDrawerSection,
   PageHeader,
+  PageHeaderProps,
   RootLayout,
   ThemeProvider,
 } from "@suankularb-components/react";
-import Favicon from "./Favicon";
 
-const Layout: FC<{ children: ReactNode }> = ({ children }) => {
+// Internal components
+import Favicon from "@/components/Favicon";
+
+const Layout: FC<{
+  children: ReactNode;
+  fab?: JSX.Element;
+  pageHeader?: Partial<PageHeaderProps>;
+}> = ({ children, fab, pageHeader }) => {
+  const router = useRouter();
   const [navOpen, setNavOpen] = useState<boolean>(false);
+
+  const getIsSelected = (pattern: RegExp) => pattern.test(router.pathname);
 
   return (
     <ThemeProvider>
       <RootLayout>
+        {/* Navigation Drawer */}
         <NavDrawer open={navOpen} onClose={() => setNavOpen(false)}>
           <NavDrawerSection
             header={
@@ -39,47 +49,65 @@ const Layout: FC<{ children: ReactNode }> = ({ children }) => {
             alt="SK Components"
           >
             <NavDrawerItem
-              icon={<MaterialIcon icon="school" />}
-              label="Learn"
-              selected
-              href="/"
+              icon={<MaterialIcon icon="touch_app" />}
+              label="Actions"
+              selected={getIsSelected(/\/actions$/)}
+              href="/actions"
               element={Link}
             />
             <NavDrawerItem
-              icon={<MaterialIcon icon="groups" />}
-              label="Class"
-              href="/"
+              icon={<MaterialIcon icon="input" />}
+              label="Input"
+              selected={getIsSelected(/\/input$/)}
+              href="/input"
               element={Link}
             />
             <NavDrawerItem
-              icon={<MaterialIcon icon="search" />}
-              label="Lookup"
-              href="/"
+              icon={<MaterialIcon icon="subdirectory_arrow_right" />}
+              label="Navigation"
+              selected={getIsSelected(/\/navigation$/)}
+              href="/navigation"
               element={Link}
             />
             <NavDrawerItem
-              icon={<MaterialIcon icon="account_circle" />}
-              label="Account"
-              metadata="7 messages"
-              href="/"
+              icon={<MaterialIcon icon="dashboard" />}
+              label="Containers"
+              selected={getIsSelected(/\/containers$/)}
+              href="/containers"
               element={Link}
             />
           </NavDrawerSection>
-          <NavDrawerSection header="M.504">
+          <NavDrawerSection header="About">
             <NavDrawerItem
-              icon={<MaterialIcon icon="group" />}
-              label="My teachers"
+              icon={<MaterialIcon icon="info" />}
+              label="About"
               href="/"
               element={Link}
             />
             <NavDrawerItem
-              icon={<MaterialIcon icon="contacts" />}
-              label="My classmates"
-              href="/"
-              element={Link}
+              icon={<MaterialIcon icon="description" />}
+              label="Documentation"
+              href="https://docs.google.com/document/d/1UJeTpXcB2MBL9Df4GUUeZ78xb-RshNIC_-LCIKmCo-8/edit?usp=sharing"
+              element={(props) => (
+                <a {...props} target="_blank" rel="noreferrer">
+                  {props.children}
+                </a>
+              )}
+            />
+            <NavDrawerItem
+              icon={<MaterialIcon icon="handshake" />}
+              label="Contribute"
+              href="https://github.com/suankularb-wittayalai-school/sk-components"
+              element={(props) => (
+                <a {...props} target="_blank" rel="noreferrer">
+                  {props.children}
+                </a>
+              )}
             />
           </NavDrawerSection>
         </NavDrawer>
+
+        {/* Navigation Bar/Rail */}
         <NavBar
           brand={
             <Image
@@ -90,69 +118,50 @@ const Layout: FC<{ children: ReactNode }> = ({ children }) => {
               alt=""
             />
           }
-          fab={
-            <FAB
-              color="tertiary"
-              size="large"
-              icon={<MaterialIcon icon="add" />}
-              alt="Add"
-              stateOnScroll="disappear"
-            />
-          }
-          end={
-            <>
-              <NavBarItem
-                icon={<MaterialIcon icon="translate" />}
-                label="เปลี่ยนภาษา"
-                tooltip="Change language / เปลี่ยนภาษา"
-                href="/"
-                element={Link}
-              />
-              <NavBarItem
-                icon={<MaterialIcon icon="logout" />}
-                label="Log out"
-                href="/"
-                element={Link}
-              />
-            </>
-          }
+          fab={fab}
           onNavToggle={() => setNavOpen(true)}
         >
           <NavBarItem
-            icon={<MaterialIcon icon="school" />}
-            label="Learn"
-            selected
-            href="#"
+            icon={<MaterialIcon icon="touch_app" />}
+            label="Actions"
+            selected={getIsSelected(/\/actions$/)}
+            href="/actions"
             element={Link}
           />
           <NavBarItem
-            icon={<MaterialIcon icon="groups" />}
-            label="Class"
-            href="#"
+            icon={<MaterialIcon icon="input" />}
+            label="Input"
+            selected={getIsSelected(/\/input$/)}
+            href="/input"
             element={Link}
           />
           <NavBarItem
-            icon={<MaterialIcon icon="search" />}
-            label="Lookup"
-            href="#"
+            icon={<MaterialIcon icon="subdirectory_arrow_right" />}
+            label="Navigation"
+            selected={getIsSelected(/\/navigation$/)}
+            href="/navigation"
             element={Link}
           />
           <NavBarItem
-            icon={<MaterialIcon icon="account_circle" />}
-            label="Account"
-            badge={7}
-            href="#"
+            icon={<MaterialIcon icon="dashboard" />}
+            label="Containers"
+            selected={getIsSelected(/\/containers$/)}
+            href="/containers"
             element={Link}
           />
         </NavBar>
+
+        {/* Page Header */}
         <PageHeader
-          title="Learn"
-          icon={<MaterialIcon icon="school" />}
+          title="SK Components Demo"
           brand={<Favicon />}
           homeURL="/"
           element={Link}
           onNavToggle={() => setNavOpen(true)}
+          {...pageHeader}
         />
+
+        {/* Content */}
         {children}
       </RootLayout>
     </ThemeProvider>
